requirements.txt
djangorestframework
asgiref==3.2.10
country-list==0.2.1
Django==3.1
pytz==2020.1
sqlparse==0.3.1
